print more info in flag usage
This commit updates flag usage strings to print the short name for the
flag (if one exists) and the default value for the flag (if one exists).

It also changes the signature of the short name of flags to (Maybe
String) to make these optional and potentially longer than a single
character.