add type validation for flags
Flags are type validated, so that if a user provides e.g. a word as an
argument to an int flag, we generate and error. Note: This requires a
future change to the parse routine since we still consume an argument
when this error is encountered (we'd otherwise mess up our counting and
process and argument twice).