program: add add-flags and flag-set?
This commit adds a convenience function for adding an array of flags to
a program instead of needing to add them individually. It also adds a
function to check if a flag was set or not (which returns a bool). If
the flag has its default value, it's treated as unset. This is useful
for checking flags in when statements as follows:

  (when (Clig.flag-set? &program "my-flag")
    (do-something))

This commit also removes some redundant white space.