clig: automatically add help flag to clig programs
This commit adds an automatic help flag to all programs defined using
clig, assigned to the options strings "--help" and "-h". There is
currently no way for users to override this behavior, but perhaps we
should support that.

When parsing options, if the help option is present, the clig program
will only print its usage string and exit.