clig: respect option terminator
This commit makes Clig respect the Posix standard non-option option
delimiter `--`. If Clig encounters this argument during parsing, the
subsequent arguments will all be added to the args array, even if they
begin with an option delimiter `-`. This is necessary, for instance, for
programs that expect negative numbers as input (e.g. -80).